Move TokenUtils to common service to avoid login filters
