+++ common-service-core/build.gradle
@@ -2,6 +2,8 @@ dependencies {
     compile(libraries.eureka_client)
     compile libraries.netflix_servo
 
+    //compile group: 'org.springframework', name: 'spring-web', version: '5.1.3.RELEASE'
+    compile(libraries.spring_boot_starter_web)
     compile(libraries.http_client)
     compile(libraries.http_core)
     compile(libraries.jackson_annotations)
