+++ zaas-client/build.gradle
@@ -1,5 +1,5 @@
 dependencies {
-    implementation project(':apiml-security-common')
+    implementation project(':apiml-common')
 
     annotationProcessor libraries.lombok
